Add debug logging, optional flag
Add `debug_flag` parameter for creating an instance of Secretbox. This
enables the class logger to `DEBUG` levels (normally `ERROR`). Any
output with values of loaded variables is truncated, only showing the
last one-fourth of the value.

Additionally, add E203 to `flake8` ignore list because we aren't going
to fight `black` for who is correct on where the `:` goes in a list
splice.