Filter response body from botocore parsers
botocore.parsers will display the entire response body in a `DEBUG`
level log. As this can include secrets/tokens that we do not want in
console output (or cloudwatch) we add a filter to `botocore.parsers` and
redact the secrets. This filter is removed in a `finally` block,
hopefully leaving the enviornment unpolluted.