Allow `boto3-stubs` to be optional for secretmanager
In the common deployment case that secretbox is being used on an AWS
Lambda for access to the secrets manager, `boto3` library is already
installed on the lambda. This change stops secretbox from defining
`boto3` as `None` in the event that only
`mypy_boto3_secretsmanager.client` is missing.

Fix test that was passing for the wrong reason, confirm it is passing
for the correct reason.