Allow `boto3-stubs` to be optional for parameter store
In the common deployment case that secretbox is being used on an AWS
Lambda for access to the parameter store, `boto3` library is already
installed on the lambda. This change stops secretbox from defining
`boto3` as `None` in the event that only
`mypy_boto3_ssm.client` is missing.

Replace `Exception` with `ClientError` for capturing AWS exceptions.

Replace error logging with `ClientError` specific logging method for
clarity in logs.

Rename client connection method to avoid `mypy` conflicts with
superclass return. Consider breaking this inheritance and moving mutual
methods to an interface for AWS.