Strip down secretbox.py for 3.0
A complete rewrite of the library from the ground up. The prior
mentality of design was for secretbox to be silent and swallow errors
over raising during run-time. It was expected that missing values or
failures would, eventually, cause an error.

The new approach will focus on a fail-fast approach.

SecretBox objects will hold a `dict[str, str]`. Optionally this
can be pre-loaded with the existing environment variables from
`os.environ`.

The SecretBox object should behave, within reason, as a dictionary. Due
to opinion, all keys in the SecretBox store will be normilized to
uppercase.

Support for `__getitem__` and `__setitem__` added.