Remove auto-load of environment variables
After consideration of #160, the loading of environment variables should
be handled exclusively in a loader. This is kicking the can on how to
handle `os.environ` behavior when in a Windows OS. However, it makes
sense that the SecretBox instance should be case sensitive with keys. It
is patterned to behave like a dictionary. Changing the behavior of
`__setitem__` and `__getitem__` here would be more surprising to a dev
than not.