Implement .get_int() method
This method attempts to return the requested value as an integer instead
of a string. Exceptions are raised if the convertion fails. All other
behavior mirrors `.get()`.

There is the question of what to do with floats here. If the value is a
float it will successfully be converted to an int. However, that changes
the value. We should likely raise an exception in this case.