Implement loader method (#172)
* Correct coverage exclusion line

There is no reason I can think of that `if not TYPE_CHECKING:` should
ever be excluded. Not sure where this came from but fixing it now as it
had an extra space causing coverage issues with this PR.

* Implement loader method

Run one or more loaders which return a `dict[str, str]`. Update the
key:pair store in the class with the loaded values.

* Handle no coverage lines

I found the reason to ignore `if not TYPE_CHECKING:`. I'm okay with the
pragma comments for now.