Implement strict behavior on key assignment (#173)
During assignment or load() actions, if a key already exists then raise
a KeyError by default. This is to prevent the clobbering of existing
key:value pairs in the SecretBox.

This behavior can be disabled with the `raise_on_overwrite` keyword
parameter during initialization. This behavior should not be changed
once the SecretBox is initialized.